# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pillars/{slug}/resources:
    get:
      operationId: getPillarResources
      tags: [sources]
      summary: Get curated external resource links for a pillar
      description: >
        Returns operator-seeded external resource links for the given pillar,
        ordered by display_order. Public, read-only. Returns an empty array for
        pillars with no resource rows. Used to surface "Learn more" panels on
        the UNCOVERED agent path.
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [resources]
                properties:
                  resources:
                    type: array
                    items: { $ref: "#/components/schemas/PillarResource" }
        "404":
          description: Pillar not found
          content:
            application/json:
              schema:
                type: object
                required: [error]
                properties:
                  error: { type: string }
...
components:
  schemas:
...
    PillarResource:
      type: object
      required: [id, title, url, displayOrder]
      properties:
        id: { type: integer }
        title: { type: string }
        url: { type: string }
        description: { type: string, nullable: true }
        category: { type: string, nullable: true }
        displayOrder: { type: integer }
...